Rename time.py top time_functions.py
- time.py shadowed pythons internal time module
