+++ src/fagfunksjoner/dapla/standards/time_functions.py
